# Use simplified `Exp` with reduced domain in Softmax op