The middle step of the softmax operation computes `exp(x - max(x))`.  Since we know the exp argument is <= 0, and this operation can tolerate a higher lower cutoff, we can use a simplified version of `Exp` with a reduced range which is
more efficient.

ONNX Runtime has a [similar optimization](https://github.com/microsoft/onnxruntime/blob/25c43c1fd66de00d3e7ebe58ecd4b37e060e5be5/onnxruntime/core/mlas/lib/compute.cpp#L312), as does [XNNPack](https://github.com/google/XNNPACK/blob/7cbb479fd61f01bc56045ac3a449274c78a68409/src/f32-raddstoreexpminusmax/avx2-rr2-p5.c.in#L18).

Tested on M3 Pro, this is approximately 20% faster.

```
 cargo test -p rten-vecmath -r -- --ignored --nocapture bench_softmax

# Before
reference 193835 us vectorized 51013 us. reference / vectorized ratio 3.800

# After
reference 194822 us vectorized 42234 us. reference / vectorized ratio 4.613
```